requirements.txt
google-genai
sarvamai
python-dotenv
fastapi
uvicorn[standard]